Replace gatsby-plugin-ipfs with modified / fixed copied code
gatsby-plugin-ipfs was/is required to make relative paths work so we can
build to Electron.

gatsby-plugin-ipfs doesn't do everything we need & breaks some paths
with its prefixing. This commit includes the original plugin code with
added fixes and modifications.
Fix for making it run on gatsby 4 has been submitted as PR, fixes for
icons/mdx-images use regex replacements that may break other projects so
they're for this project only atm.